% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H5S_Array.R, R/RHDF5Array.R
\name{as}
\alias{as}
\title{coercion for remote array to remote matrix}
\description{
coercion for remote array to remote matrix

Coercion method from HSDSMatrix to its superclass HSDSArray
}
\seealso{
Other HSDSArray: 
\code{\link{HSDSArray}},
\code{\link{HSDSMatrix}}
}
\concept{H5S_Array}
\concept{HSDSArray}
