% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R
\name{dataset}
\alias{dataset}
\title{Find a dataset on source from its name}
\usage{
dataset(h5s, tag)
}
\arguments{
\item{h5s}{instance of H5S_source}

\item{tag}{character string identifying a dataset}
}
\value{
object of type H5S_dataset
}
\description{
This function is deprecated and will be defunct in the next release.
}
