% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H5S_Array.R
\name{dim,H5S_ArraySeed-method}
\alias{dim,H5S_ArraySeed-method}
\title{HDF Server content is assumed transposed relative to R matrix layout}
\usage{
\S4method{dim}{H5S_ArraySeed}(x)
}
\arguments{
\item{x}{instance of H5S_ArraySeed}
}
\value{
integer(2) vector of dimensions corresponding to R's layout, assuming 2-d data
}
\description{
This function is deprecated and will be defunct in the next release.
}
