% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H5S_Array.R, R/RHDF5Array.R
\docType{methods}
\name{dim}
\alias{dim}
\alias{dim,H5S_ArraySeed-method}
\alias{dim,HSDSArraySeed-method}
\title{HDF Server content is assumed transposed relative to R matrix layout}
\usage{
\S4method{dim}{H5S_ArraySeed}(x)

\S4method{dim}{HSDSArraySeed}(x)
}
\arguments{
\item{x}{instance of H5S_ArraySeed}

\item{x}{An object of type HSDSArraySeed}
}
\value{
integer(2) vector of dimensions corresponding to R's layout, assuming 2-d data

A numeric vector of the dimensions
}
\description{
This function is deprecated and will be defunct in the next release.

(required by DelayedArray seed contract)
HDF server content is assumed transposed relative to R matrix layout.
This anticipates H5 datasets on the server with rows for 
experimental samples and columns for *-omic features. The 
Bioconductor SummarizedExperiment requires *-omic features in 
rows and samples in columns.
}
