% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H5S_Array.R, R/RHDF5Array.R
\docType{methods}
\name{dimnames,H5S_ArraySeed-method}
\alias{dimnames,H5S_ArraySeed-method}
\alias{dimnames}
\alias{dimnames,HSDSArraySeed-method}
\title{dimnames not stored with H5S_source as of Jan 2018}
\usage{
\S4method{dimnames}{H5S_ArraySeed}(x)

\S4method{dimnames}{HSDSArraySeed}(x)
}
\arguments{
\item{x}{An object of type HSDSArraySeed or (deprecated) H5S_ArraySeed}
}
\value{
currently returns list(NULL, NULL) as we do not store dimnames in HDF5

A NULL list of length equal to the array dimensionality
}
\description{
This function is deprecated and will be defunct in the next release.

(required by DelayedArray seed contract, returns NULL list)
}
