% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5servhsds.R
\name{domains}
\alias{domains}
\title{HSDS server domains accessor}
\usage{
domains(object)
}
\arguments{
\item{object}{H5S_source instance}

\item{\dots}{not used}
}
\value{
a data frame with domains name
}
\description{
HSDS server domains accessor
}
