% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R
\docType{methods}
\name{domains}
\alias{domains}
\alias{domains,H5S_source-method}
\title{HSDS server domains accessor}
\usage{
domains(object, ...)
}
\arguments{
\item{object}{H5S_source instance}

\item{\dots}{not used}
}
\value{
a data frame with domains name
}
\description{
HSDS server domains accessor
}
\examples{
hsdsCon = H5S_source(URL_hsds()) # hsds server connection
setPath(hsdsCon, "/home/stvjc/")-> hsds
domains(hsds)
}
