% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R
\name{dsmeta}
\alias{dsmeta}
\title{list information about datasets available in an H5S_source}
\usage{
dsmeta(src)
}
\arguments{
\item{src}{H5S_source instance}
}
\value{
data frame with one row for each group and three columns. The 
second column has the list of datasets in the group.
}
\description{
list information about datasets available in an H5S_source
}
\examples{
bigec2 = H5S_source("http://h5s.channingremotedata.org:5000")
dsm <- dsmeta(bigec2) 
dst <- unlist(dsm[1,2])    # all dataset candidates in group 1
}
