% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R, R/Dataset.R
\docType{methods}
\name{[,H5S_dataset,numeric,numeric-method}
\alias{[,H5S_dataset,numeric,numeric-method}
\alias{[,H5S_dataset,character,character-method}
\alias{[}
\alias{[,HSDSDataset-method}
\alias{[,HSDSDataset,numeric-method}
\alias{[,HSDSDataset,numeric,numeric-method}
\alias{[,HSDSDataset,numeric,ANY-method}
\alias{[,HSDSDataset,numeric,ANY,ANY-method}
\title{extract elements from H5S_dataset}
\usage{
\S4method{[}{H5S_dataset,numeric,numeric}(x, i, j, ..., drop = FALSE)

\S4method{[}{H5S_dataset,character,character}(x, i, j, ..., drop = FALSE)

\S4method{[}{HSDSDataset,numeric,ANY}(x, i)
}
\arguments{
\item{x}{instance of H5S_dataset}

\item{i}{select option for first matrix index in HDF5 server value API}

\item{j}{select option for second matrix index in HDF5 server value API}

\item{\dots}{unused}

\item{drop}{logical defaults to FALSE}

\item{x}{object of type HSDSDataset}

\item{i}{vector of indices (first dimension)}
}
\value{
matrix of data obtained

an array with the elements requested from the HSDSDataset
}
\description{
This function is deprecated and will be defunct in the next release.

This function is deprecated and will be defunct in the next release.
}
