% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{extractCompoundJSON}
\alias{extractCompoundJSON}
\title{compound operation}
\usage{
extractCompoundJSON(type, value)
}
\arguments{
\item{type}{type}

\item{value}{value}
}
\description{
compound operation
}
