% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R
\name{fetchDatasets}
\alias{fetchDatasets}
\title{fetch datasets of a hdf5 file from the hsds server}
\usage{
fetchDatasets(object)
}
\arguments{
\item{object}{instance of H5S_source}
}
\description{
fetch datasets of a hdf5 file from the hsds server
}
\examples{
hsdsCon = H5S_source(URL_hsds()) # hsds server
hsdsCon@folderPath="/home/stvjc/hdf5_mat.h5"
ds = fetchDatasets(hsdsCon)
ds$datasets
}
