% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5servhsds.R
\name{getDatasetAttrs}
\alias{getDatasetAttrs}
\title{getDatasetAttrs from hsds server}
\usage{
getDatasetAttrs(object)
}
\arguments{
\item{object}{instance of H5S_source(updated object with credentials set)}
}
\description{
getDatasetAttrs from hsds server
}
