% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R
\name{getDatasetSlice}
\alias{getDatasetSlice}
\title{getDatasetSlice from hsds server}
\usage{
getDatasetSlice(object, dsindex = 1, selectionString, ...)
}
\arguments{
\item{object}{instance of H5S_source(updated object with path to file set)}

\item{dsindex}{dataset index}

\item{\dots}{unused}

\item{selectionsString}{character with selectionString}
}
\value{
list of data obtained
}
\description{
getDatasetSlice from hsds server
}
\examples{
hsdsCon = H5S_source("http://149.165.156.12:5101") # hsds server
setPath(hsdsCon, "/home/reshg/tenx_full2.h5")-> hsds
getDatasetSlice(hsds,dsindex=1,selectionString="[1:10,1:50]")
}
