% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File.R
\name{getHSDSFileList}
\alias{getHSDSFileList}
\title{List files on a directory on HSDS-server}
\usage{
getHSDSFileList(url, directory = NULL)
}
\arguments{
\item{url, }{containing url of the server and root domain.}

\item{directory, }{containing name of the directory}
}
\value{
List of all HDF5-files on the server or all files of the collection
}
\description{
This funstions returns a list of all HDF5-files on the server or all files of the collection
}
\examples{
if (check_hsds()) {
 url <- 'https://alserglab.wustl.edu/hsds/?domain=/counts'
 getHSDSFileList(url)
}
}
