% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R
\name{getReq}
\alias{getReq}
\title{list information about server content available in an H5S_source hsds instance}
\usage{
getReq(src)
}
\arguments{
\item{src}{H5S_source instance}
}
\value{
data frame with 5 columns for one row for each user's data
}
\description{
list information about server content available in an H5S_source hsds instance
}
