% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R
\docType{class}
\name{H5S_dataset-class}
\alias{H5S_dataset-class}
\alias{[,H5S_dataset,numeric,numeric-method}
\title{name H5S_dataset
rdname H5S_dataset-class}
\usage{
\S4method{[}{H5S_dataset,numeric,numeric}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{instance of H5S_dataset}

\item{i}{character string usable as select option for first matrix index in HDF5 server value API}

\item{j}{character string usable as select option for second matrix index in HDF5 server value API}

\item{\dots}{unused}

\item{drop}{logical defaults to FALSE}
}
\value{
matrix of data obtained
}
\description{
name H5S_dataset
rdname H5S_dataset-class

extract elements from H5S_dataset
}
\section{Slots}{

\describe{
\item{\code{source}}{instance of H5S_source instance}

\item{\code{simpleName}}{character string naming dataset}

\item{\code{shapes}}{list including dimension information}

\item{\code{hrefs}}{DataFrame of hrefs as defined in the API}

\item{\code{allatts}}{list of all attributes}

\item{\code{presel}}{string prepared for select operation in GET}

\item{\code{transfermode}}{default "JSON" or "binary" for binary transfer}
}}

