% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R
\docType{class}
\name{H5S_dataset}
\alias{H5S_dataset}
\alias{H5S_dataset-class}
\title{construct H5S_dataset object}
\description{
construct H5S_dataset object
}
\section{Slots}{

\describe{
\item{\code{source}}{instance of H5S_source instance}

\item{\code{simpleName}}{character string naming dataset}

\item{\code{shapes}}{list including dimension information}

\item{\code{hrefs}}{DataFrame of hrefs as defined in the API}

\item{\code{allatts}}{list of all attributes}

\item{\code{presel}}{string prepared for select operation in GET}

\item{\code{transfermode}}{default "JSON" or "binary" for binary transfer}
}}

