% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5servhsds.R
\name{H5S_dataset2}
\alias{H5S_dataset2}
\title{H5S_dataset2 for datasets in hsds server}
\usage{
H5S_dataset2(object)
}
\arguments{
\item{object}{instance of H5S_source(updated object with credentials set)}
}
\description{
H5S_dataset2 for datasets in hsds server
}
