% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R
\docType{class}
\name{H5S_source}
\alias{H5S_source}
\alias{H5S_source-class}
\alias{[[,H5S_source,character-method}
\title{H5S_source identifies an HDF5/HSDS server and manages some metadata about contents}
\usage{
H5S_source(serverURL, domain, ...)

\S4method{[[}{H5S_source,character}(x, i, j)
}
\arguments{
\item{serverURL}{a URL for a port for HDF5Server}

\item{domain}{character string with path to file for HSDS}

\item{\dots}{not used}

\item{x}{instance of H5S_source}

\item{i}{character string intended to identify dataset on server}

\item{j}{not used}
}
\value{
an initialized object of type H5S_source
}
\description{
H5S_source identifies an HDF5/HSDS server and manages some metadata about contents

construct H5S_source
}
\section{Slots}{

\describe{
\item{\code{serverURL}}{character string with a URL}

\item{\code{dsmeta}}{DataFrame instance with metadata about content of h5serv server}

\item{\code{dmains}}{DataFrame instance with metadata about the content of hsds server}

\item{\code{getReq}}{DataFrame instance with metadata about hsds server}

\item{\code{folderPath}}{character string with path to user's folder/file on hsds server}
}}

\note{
The dsmeta slot holds a DataFrame with a column \code{dsnames}
that is a list with ith element a character vector of all dsnames
available for the ith group.  There is no effort at present to
search all groups for candidate datasets.

If the domain for the HSDS server is known, 
pass the domain path as a character string along with ther serverURL
}
\examples{
\dontrun{
bigec2 = H5S_source(URL_h5serv()) # h5serv 
bigec2
dsmeta(bigec2)[1:2,]       # two groups
dsmeta(bigec2)[1,2][[1]]   # all dataset candidates in group 1
}
hsdsCon = H5S_source(URL_hsds()) # hsds server connection
hsdsCon
getReq(hsdsCon)
setPath(hsdsCon,"/home/stvjc/hdf5_mat.h5") -> hsds
fetchDatasets(hsds)     # grab the dataset id of interest 
H5S_dataset2(hsds, "d-a9e4b71c-8ea2-11e8-9306-0242ac120022")
}
