% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5servhsds.R
\name{hsdel}
\alias{hsdel}
\title{hsdel - Delete hdf5 file on hsds server}
\usage{
hsdel(object, domain)
}
\arguments{
\item{object}{instance of H5S_source (updated object with credentials set)}

\item{domain}{HDF Server domain (Unix or DNS style) - path to hdf5 file to be deleted}
}
\description{
hsdel - Delete hdf5 file on hsds server
}
\examples{
require(rhdf5client)
bigec2 = H5S_source("http://149.165.156.12:5101") # a hsds server
setCredentials(bigec2,username="username",password="password",folderPath="/home/reshg/") -> new_bigec2
hsdel(new_bigec2, domain)
}
