% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSDS_Matrix.R
\name{HSDS_Matrix}
\alias{HSDS_Matrix}
\title{simplify construction of DelayedMatrix from url and path in HSDS}
\usage{
HSDS_Matrix(url, path)
}
\arguments{
\item{url}{character(1) URL for HSDS object store with port}

\item{path}{character(1) path from root defining HDF Cloud resource}
}
\value{
instance of DelayedArray
}
\description{
simplify construction of DelayedMatrix from url and path in HSDS
}
\examples{
HSDS_Matrix
}
