% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHDF5Array.R
\docType{class}
\name{HSDSArraySeed}
\alias{HSDSArraySeed}
\alias{HSDSArraySeed-class}
\title{HSDSArraySeed for HSDSArray backend to DelayedArray}
\usage{
HSDSArraySeed(endpoint, svrtype, domain, dsetname)
}
\arguments{
\item{endpoint}{URL of remote server}

\item{svrtype}{type of server, must be either 'hsds' or 'h5serv'}

\item{domain}{HDF5 domain of H5 file on server}

\item{dsetname}{complete internal path to dataset in H5 file}
}
\value{
An initialized object of type HSDSArraySeed
}
\description{
HSDSArraySeed for HSDSArray backend to DelayedArray

Construct an object of type HSDSArraySeed
}
\section{Slots}{

\describe{
\item{\code{endpoint}}{URL of remote server}

\item{\code{svrtype}}{type of server, must be either 'hsds' or 'h5serv'}

\item{\code{domain}}{HDF5 domain of H5 file on server}

\item{\code{dsetname}}{complete internal path to dataset in H5 file}

\item{\code{dataset}}{object of type HSDSDataset for access to the H5 dataset}
}}

