% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\docType{class}
\name{HSDSDataset-class}
\alias{HSDSDataset-class}
\title{An S4 class to represent a dataset in a HDF5 file.}
\description{
An S4 class to represent a dataset in a HDF5 file.
}
\section{Slots}{

\describe{
\item{\code{file}}{An object of type HSDSFile; the file in which the dataset is resident.}

\item{\code{path}}{The dataset's path in the internal HDF5 hiearchy.}

\item{\code{uuid}}{The unique unit ID by which the dataset is accessed in the server 
database system.}

\item{\code{shape}}{The dimensions of the dataset}

\item{\code{type}}{The dataset's HDF5 datatype}
}}

