% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{HSDSDataset}
\alias{HSDSDataset}
\title{Construct an object of type HSDSDataset}
\usage{
HSDSDataset(file, path)
}
\arguments{
\item{file}{An object of type HSDSFile which hosts the dataset}

\item{path}{The complete intrafile path to the dataset}
}
\description{
A HSDSDataset is a representation of a dataset in a HDF5 file.
}
\examples{
src <- HSDSSource('http://hsdshdflab.hdfgroup.org')
f <- HSDSFile(src, '/home/spollack/testzero.h5')
d <- HSDSDataset(f, '/grpA/grpAB/dsetX')
}
