% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5servhsds.R
\name{hsdsInfo}
\alias{hsdsInfo}
\title{HSDS server get request accessor}
\usage{
hsdsInfo(object)
}
\arguments{
\item{object}{H5S_source instance}

\item{\dots}{not used}
}
\value{
a data frame with response
}
\description{
HSDS server get request accessor
}
