% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R
\docType{methods}
\name{hsdsInfo}
\alias{hsdsInfo}
\alias{hsdsInfo,H5S_source-method}
\title{HSDS server get request accessor}
\usage{
hsdsInfo(object)
}
\arguments{
\item{object}{H5S_source instance}
}
\value{
a data frame with response
}
\description{
HSDS server get request accessor
}
\examples{
hsdsCon = H5S_source("http://149.165.156.12:5101") # hsds server connection
hsdsInfo(hsdsCon)
}
