% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R
\docType{methods}
\name{hsdsInfo}
\alias{hsdsInfo}
\alias{hsdsInfo,H5S_source-method}
\title{HSDS server get request accessor}
\usage{
hsdsInfo(object)
}
\arguments{
\item{object}{H5S_source instance}
}
\value{
a data frame with response
}
\description{
This function is deprecated and will be defunct in the next release.
}
\examples{
hsdsCon = H5S_source(URL_hsds()) # hsds server connection
hsdsInfo(hsdsCon)
}
