% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHDF5Array.R
\docType{class}
\name{HSDSMatrix}
\alias{HSDSMatrix}
\alias{HSDSMatrix-class}
\title{DelayedMatrix subclass for a two-dimensional HSDSArray}
\description{
DelayedMatrix subclass for a two-dimensional HSDSArray
}
\seealso{
Other HSDSArray: \code{\link{HSDSArray}}, \code{\link{as}}
}
\concept{HSDSArray}
