% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source.R
\docType{class}
\name{HSDSSource-class}
\alias{HSDSSource-class}
\title{An S4 class to represent a HDF5 server listening on a port.}
\description{
An S4 class to represent a HDF5 server listening on a port.
}
\section{Slots}{

\describe{
\item{\code{endpoint}}{URL for server}

\item{\code{type}}{Type of server software at the source; must be 
either 'h5serv' or (default) 'hsds'}
}}

