% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5servhsds.R
\name{hsget}
\alias{hsget}
\title{hsget - Copy server domain to local HDF5 file( get content from HSDS server to local destination)}
\usage{
hsget(object, domain, des)
}
\arguments{
\item{object}{instance of H5S_source (updated object with credentials set)}

\item{domain}{HDF Server domain (Unix or DNS style) - path to hdf5 file to be copied}

\item{des}{local destination path where the hsds hdf5 file should be copied to}
}
\description{
hsget - Copy server domain to local HDF5 file( get content from HSDS server to local destination)
}
\examples{
require(rhdf5client)
bigec2 = H5S_source("http://149.165.156.12:5101") # a hsds server
setCredentials(bigec2,username="username",password="password",folderPath="/home/reshg/") -> new_bigec2
hsget(new_bigec2, domain, des)
}
