% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5servhsds.R
\name{hsload}
\alias{hsload}
\title{hsload - Load datasets to hsds server (hsload version 0.0.1) / Copy HDF5 file to Domain or multiple files to a Domain folder}
\usage{
hsload(object, source, domain)
}
\arguments{
\item{object}{instance of H5S_source(updated object with credentials set)}

\item{source}{HDF5 file or multiple files if copying to folder - source file path/name}

\item{domain}{HDF Server domain (Unix or DNS style) - destination path/name}
}
\description{
hsload - Load datasets to hsds server (hsload version 0.0.1) / Copy HDF5 file to Domain or multiple files to a Domain folder
}
\examples{
require(rhdf5client)
bigec2 = H5S_source("http://149.165.156.12:5101") # a hsds server
setCredentials(bigec2,username="username",password="password",folderPath="/home/reshg/") -> new_bigec2
hsload(new_bigec2, source, domain)
}
