% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5servhsds.R
\name{hsls}
\alias{hsls}
\title{hsls - View datasets on hsds server}
\usage{
hsls(object)
}
\arguments{
\item{object}{instance of H5S_source(updated object with credentials set)}
}
\description{
hsls - View datasets on hsds server
}
\examples{
require(rhdf5client)
bigec2 = H5S_source("http://149.165.156.12:5101") # a hsds server
setCredentials(bigec2,username="username",password="password",folderPath="/home/reshg/") -> new_bigec2
hsls(new_bigec2)
}
