% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5servhsds.R
\name{hstouch}
\alias{hstouch}
\title{hstouch - Create a domain on HSDS server}
\usage{
hstouch(object, domain)
}
\arguments{
\item{object}{instance of H5S_source (updated object with credentials set)}

\item{domain}{HDF Server domain (Unix or DNS style) - domain to be created}
}
\description{
hstouch - Create a domain on HSDS server
}
\examples{
require(rhdf5client)
bigec2 = H5S_source("http://149.165.156.12:5101") # a hsds server
setCredentials(bigec2,username="username",password="password",folderPath="/home/reshg/") -> new_bigec2
hstouch(new_bigec2, domain)
}
