% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R
\name{internalDim}
\alias{internalDim}
\title{acquire internal HDF5 dimension information for matrix}
\usage{
internalDim(h5d)
}
\arguments{
\item{h5d}{instance of H5S_dataset}
}
\value{
vector with dimensions of dataset
}
\description{
acquire internal HDF5 dimension information for matrix
}
\examples{
\dontrun{
bigec2 = H5S_source(URL_h5serv())
tex <- bigec2[["tenx_100k_sorted"]]
internalDim(tex)
}
}
