% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source.R
\docType{methods}
\name{listDomains}
\alias{listDomains}
\alias{listDomains,HSDSSource,character-method}
\alias{listDomains,HSDSSource,missing-method}
\title{List files and subdirectories of a domain}
\usage{
listDomains(object, rootdir)

\S4method{listDomains}{HSDSSource,character}(object, rootdir)

\S4method{listDomains}{HSDSSource,missing}(object)
}
\arguments{
\item{object}{An object of type HSDSSource}

\item{rootdir}{A slash-separated directory in the HSDSSource file system.}
}
\value{
a vector of domains in the rootdir
}
\description{
The user needs to give the domain to start in. The search
will be non-recursive. I.e., output for domain '/home/jreadey/' will 
not return the files in '/home/jreadey/HDFLabTutorial/'
}
\examples{
src.hsds <- HSDSSource('http://hsdshdflab.hdfgroup.org')
src.chan <- HSDSSource('http://h5s.channingremotedata.org:5000', 'h5serv')
listDomains(src.chan)
listDomains(src.hsds, '/home/jreadey')
}
