\name{rhdf5client-deprecated}
\alias{rhdf5client-deprecated}
\title{Deprecated functions in package \sQuote{rhdf5client}}

\description{
  These functions are provided for compatibility with older versions
  of \sQuote{rhdf5client} only, and will be defunct at the next release.
}

\details{

  The following functions are deprecated and will be made defunct in the next release: 
  \itemize{
    \item{URL_h5serv} 
    \item{URL_hsds}
    \item{dsmeta}
    \item{getReq}
    \item{groups}
    \item{setPath}
    \item{links}
    \item{transfermode}
    \item{dataset}
    \item{internalDim}
    \item{hsdsInfo}
    \item{domains}
    \item{getDatasetUUIDs}
    \item{getDatasetAttrs}
    \item{getDims}
    \item{getHRDF}
    \item{H5S_dataset2}
    \item{getDatasetSlice}
    \item{fetchDatasets}
    \item{isplit}
    \item{sproc}
    \item{listDomains}
    \item{listDatasets}
    \item{getData}
% how should these be declared?
%   \item{subscript}
%   \item{double-subscript}
%   \item{transfermode<-}
  }
  The following classes are deprecated and will be made defunct in the next release: 
  \itemize{
    \item{H5S_source}
    \item{H5S_dataset}
    \item{H5S_Array}
    \item{H5S_Matrix}
    \item{HSDSSource}
    \item{HSDSFile}
    \item{HSDSDataset}
  }
}

