% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhdf5client2.R
\docType{package}
\name{rhdf5client}
\alias{rhdf5client}
\alias{rhdf5client-package}
\title{rhdf5client: A package for accessing HDFGroup HDF5 servers from R.}
\description{
The rhdf5client package provides read-only access to HDF5 files
maintained on a server. The HDFGroup provides two servers, an
obsolescent one called `h5serv` and the newer prototype called `hsds`.
}
