% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5servhsds.R
\name{setCredentials}
\alias{setCredentials}
\title{set credentials for H5S_source instance}
\usage{
setCredentials(object, username, password, folderPath)
}
\arguments{
\item{object}{instance of H5S_source}

\item{username}{username on hsds server}

\item{password}{password on hsds server}

\item{folderPath}{path to user's folder on hsds server}
}
\value{
updated object of type H5S_source
}
\description{
set credentials for H5S_source instance
}
\examples{
require(rhdf5client)
bigec2 = H5S_source("http://149.165.156.12:5101") # a hsds server
setCredentials(bigec2,username="username",password="password",folderPath="/home/reshg/") -> new_bigec2
new_bigec2
}
