% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5servhsds.R
\docType{methods}
\name{setPath}
\alias{setPath}
\alias{setPath,H5S_source,character-method}
\title{set path for hsds server resource}
\usage{
setPath(object, folderPath, ...)
}
\arguments{
\item{object}{H5S_source instance}

\item{folderPath}{character string with path to user's folder on hsds server}

\item{\dots}{not used}
}
\value{
an updated object with folderPath set
}
\description{
set path for hsds server resource
}
\examples{
hsdsCon = H5S_source("http://149.165.156.12:5101") # hsds server connection
setPath(hsdsCon, "/home/reshg/tenx_full2.h5")-> hsds
}
