% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5servhsds.R
\name{setPath}
\alias{setPath}
\alias{setPath,H5S_source,missing-method}
\title{set path for hsds server resource}
\usage{
setPath(object, folderPath, ...)
}
\arguments{
\item{object}{H5S_source instance}

\item{folderPath}{character string with path to user's folder on hsds server}
}
\description{
set path for hsds server resource
}
