% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indx.R
\name{isplit}
\alias{isplit}
\alias{sproc}
\title{This function is deprecated and will be defunct in the next release.}
\usage{
isplit(x)

sproc(spl)
}
\arguments{
\item{x}{a numeric vector (should be integers)}

\item{spl}{output of isplit}
}
\value{
list of vectors of integers which can be expressed as initial/final/stride triplets

list of colon-delimited strings each with initial/final/stride triplet
}
\description{
isplit converts a numeric vector into a list of sequences for compact reexpression
}
\examples{
inds = c(1:10, seq(25,50,2), seq(200,150,-2))
sproc(isplit(inds))
}
