% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R
\docType{methods}
\name{[,H5S_dataset,character,character,ANY-method}
\alias{[,H5S_dataset,character,character,ANY-method}
\alias{[,H5S_dataset,character,character-method}
\title{extract elements from H5S_dataset}
\usage{
\S4method{[}{H5S_dataset,character,character,ANY}(x, i, j, ...,
  drop = FALSE)
}
\arguments{
\item{x}{instance of H5S_dataset}

\item{i}{character vector of row selections}

\item{j}{character vector of column selections}

\item{\dots}{not used}

\item{drop}{logical(1) set TRUE to drop array character}
}
\description{
extract elements from H5S_dataset
}
