% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R
\name{[,H5S_dataset,numeric,numeric,ANY-method}
\alias{[,H5S_dataset,numeric,numeric,ANY-method}
\title{extract elements from H5S_dataset}
\usage{
\S4method{[}{H5S_dataset,numeric,numeric,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{instance of H5S_dataset}

\item{i}{select option for first matrix index in HDF5 server value API}

\item{j}{select option for second matrix index in HDF5 server value API}

\item{\dots}{unused}

\item{drop}{logical defaults to FALSE}
}
\value{
matrix of data obtained
}
\description{
This function is deprecated and will be defunct in the next release.
}
