% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\docType{methods}
\name{[,HSDSDataset,numeric,ANY,ANY-method}
\alias{[,HSDSDataset,numeric,ANY,ANY-method}
\title{bracket method for 1d request from HSDSDataset}
\usage{
\S4method{[}{HSDSDataset,numeric,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{object of type HSDSDataset}

\item{i}{vector of indices (first dimension)}

\item{j}{not used}

\item{\dots}{not used}

\item{drop}{logical(1) if TRUE return has no array character}
}
\value{
an array with the elements requested from the HSDSDataset
}
\description{
bracket method for 1d request from HSDSDataset
}
