% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R
\name{[[}
\alias{[[}
\alias{[[,H5S_source,character-method}
\title{Subscript operator}
\usage{
\S4method{[[}{H5S_source,character}(x, i, j)
}
\arguments{
\item{x}{instance of H5S_source}

\item{i}{character string intended to identify dataset on server}

\item{j}{not used}
}
\description{
This function is deprecated and will be defunct in the next release.
}
