% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submatrix.R
\name{submatrix}
\alias{submatrix}
\title{Utility function for retrieving matrices from a two-dimensional dataset}
\usage{
submatrix(dset, i, j)
}
\arguments{
\item{dset}{A two-dimensional dataset of type H5S_dataset}

\item{i}{A vector of row indices to retrieve}

\item{j}{A vector of column indices to retrieve}
}
\value{
The submatrix defined by the indices from the dataset
}
\description{
Utility function for retrieving matrices from a two-dimensional dataset
}
