% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R
\docType{methods}
\name{transfermode<-}
\alias{transfermode<-}
\alias{transfermode<-,H5S_dataset-method}
\title{replace transfer mode}
\usage{
transfermode(object) <- value
}
\arguments{
\item{object}{instance of H5S_linkset}

\item{value}{either "JSON" (default) or "binary"}
}
\value{
updated object of type H5S_dataset
}
\description{
This function is deprecated and will be defunct in the next release.
}
