% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R
\name{URL_hsds}
\alias{URL_hsds}
\title{manage hsds URL}
\usage{
URL_hsds()
}
\value{
URL of hsds server
}
\description{
This function is deprecated and will be defunct in the next release.
}
\examples{
URL_hsds()
}
