% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHDF5Array.R
\name{as}
\alias{as}
\title{Coercion method from HSDSMatrix to its superclass HSDSArray}
\description{
Coercion method from HSDSMatrix to its superclass HSDSArray
}
\seealso{
Other HSDSArray: \code{\link{HSDSArray}},
  \code{\link{HSDSMatrix}}
}
\concept{HSDSArray}
