% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHDF5Array.R
\docType{methods}
\name{dim}
\alias{dim}
\alias{dim,HSDSArraySeed-method}
\title{Obtain dimensions of an object of type HSDSArraySeed}
\usage{
\S4method{dim}{HSDSArraySeed}(x)
}
\arguments{
\item{x}{An object of type HSDSArraySeed}
}
\value{
A numeric vector of the dimensions
}
\description{
(required by DelayedArray seed contract)
HDF server content is assumed transposed relative to R matrix layout.
This anticipates H5 datasets on the server with rows for 
experimental samples and columns for *-omic features. The 
Bioconductor SummarizedExperiment requires *-omic features in 
rows and samples in columns.
}
