% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHDF5Array.R
\docType{methods}
\name{dimnames}
\alias{dimnames}
\alias{dimnames,HSDSArraySeed-method}
\title{Obtain names of dimensions for an object of type HSDSArraySeed}
\usage{
\S4method{dimnames}{HSDSArraySeed}(x)
}
\arguments{
\item{x}{An object of type HSDSArraySeed}
}
\value{
A NULL list of length equal to the array dimensionality
}
\description{
(required by DelayedArray seed contract, returns NULL list)
}
