% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R
\name{getDatasetAttrs}
\alias{getDatasetAttrs}
\title{getDatasetAttrs from hsds server}
\usage{
getDatasetAttrs(object)
}
\arguments{
\item{object}{instance of H5S_source(updated object with path to file set)}
}
\value{
list of data obtained
}
\description{
getDatasetAttrs from hsds server
}
\examples{
hsdsCon = H5S_source(URL_hsds()) # hsds server
setPath(hsdsCon, "/home/stvjc/tenx_full.h5")-> hsds
getDatasetAttrs(hsds)
}
