% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5servhsds.R
\name{getDatasetSlice}
\alias{getDatasetSlice}
\title{getDatasetSlice from hsds server}
\usage{
getDatasetSlice(object, dsindex = 1, selectionString, ...)
}
\arguments{
\item{object}{instance of H5S_source(updated object with credentials set)}

\item{dsindex}{dataset index}

\item{selectionsString}{character with selectionString}
}
\description{
getDatasetSlice from hsds server
}
