% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5servhsds.R
\name{getDatasetUUIDs}
\alias{getDatasetUUIDs}
\title{getDatasetUUIDs from hsds server}
\usage{
getDatasetUUIDs(object)
}
\arguments{
\item{object}{instance of H5S_source(updated object with credentials set)}
}
\description{
getDatasetUUIDs from hsds server
}
