% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R
\name{getDims}
\alias{getDims}
\title{getDims from hsds server}
\usage{
getDims(object, duid)
}
\arguments{
\item{object}{instance of H5S_source(updated object with path to file set)}

\item{duid}{character string with dataset uuid}
}
\value{
numeric content of dimensions
}
\description{
This function is deprecated and will be defunct in the next release.
}
\examples{
\dontrun{
hsdsCon = H5S_source(URL_hsds()) # hsds server
setPath(hsdsCon, "/home/stvjc/hdf5_mat.h5")-> hsds
duid <- 'd-a9e4b71c-8ea2-11e8-9306-0242ac120022'
getDims(hsds, duid)
}
}
