% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R
\name{getDims}
\alias{getDims}
\title{getDims from hsds server}
\usage{
getDims(object)
}
\arguments{
\item{object}{instance of H5S_source(updated object with path to file set)}
}
\value{
numeric content of dimensions
}
\description{
getDims from hsds server
}
\examples{
hsdsCon = H5S_source(URL_hsds()) # hsds server
setPath(hsdsCon, "/home/stvjc/tenx_full.h5")-> hsds
getDims(hsds)
}
