% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5servhsds.R
\name{getDims}
\alias{getDims}
\title{getDims from hsds server}
\usage{
getDims(object)
}
\arguments{
\item{object}{instance of H5S_source(updated object with credentials set)}
}
\description{
getDims from hsds server
}
