% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5servhsds.R
\name{getHRDF}
\alias{getHRDF}
\title{getHRDF from hsds server}
\usage{
getHRDF(object)
}
\arguments{
\item{object}{instance of H5S_source(updated object with credentials set)}
}
\description{
getHRDF from hsds server
}
