% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R
\name{getHRDF}
\alias{getHRDF}
\title{getHRDF from hsds server}
\usage{
getHRDF(object)
}
\arguments{
\item{object}{instance of H5S_source(updated object with path to file set)}
}
\value{
DataFrame of data obtained
}
\description{
getHRDF from hsds server
}
\examples{
hsdsCon = H5S_source(URL_hsds()) # hsds server
setPath(hsdsCon, "/home/stvjc/tenx_full.h5")-> hsds
getHRDF(hsds)
}
