% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R
\name{H5S_dataset2}
\alias{H5S_dataset2}
\title{H5S_dataset2 for datasets in hsds server}
\usage{
H5S_dataset2(object)
}
\arguments{
\item{object}{instance of H5S_source(updated object with path to file set)}
}
\value{
H5S_dataset object
}
\description{
H5S_dataset2 for datasets in hsds server
}
\examples{
hsdsCon = H5S_source("http://149.165.156.12:5101") # hsds server
setPath(hsdsCon, "/home/reshg/tenx_full2.h5")-> hsds
H5S_dataset2(hsds)
}
