% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R
\name{H5S_dataset2}
\alias{H5S_dataset2}
\title{H5S_dataset2 for datasets in hsds server}
\usage{
H5S_dataset2(object, duid)
}
\arguments{
\item{object}{instance of H5S_source(updated object with path to file set)}

\item{duid}{character vector with dataset uuid of interest}
}
\value{
H5S_dataset object
}
\description{
H5S_dataset2 for datasets in hsds server
}
\examples{
hsdsCon = H5S_source(URL_hsds()) # hsds server
hsdsCon@folderPath="/home/stvjc/hdf5_mat.h5"
ds = fetchDatasets(hsdsCon) #Pick the dataset id of interest
H5S_dataset2(hsdsCon, "d-a9e4b71c-8ea2-11e8-9306-0242ac120022")
}
