% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source.R
\name{HSDSSource}
\alias{HSDSSource}
\title{Construct an object of type HSDSSource.}
\usage{
HSDSSource(endpoint, type = "hsds")
}
\arguments{
\item{endpoint}{URL for server}

\item{type}{Type of server software at the source; must be}
}
\value{
An object of type HSDSSource
}
\description{
A HSDSSource is a representation of a URL which provides access to a HDF5 
server (either h5serv or hsds.)
}
\examples{
src.hsds <- HSDSSource('http://hsdshdflab.hdfgroup.org')
}
