% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R
\name{links}
\alias{links}
\alias{links,H5S_source,numeric-method}
\title{access for link metadata for HDF5 server groups}
\usage{
links(object, index, ...)
}
\arguments{
\item{object}{H5S_source instance}

\item{index}{numeric group index}

\item{\dots}{not used}
}
\value{
an object of type H5S_linkset with the linkset of the group
}
\description{
access for link metadata for HDF5 server groups
}
\examples{
\dontrun{
bigec2 = H5S_source(URL_h5serv())
lks <- links(bigec2, 1)    # linkset for root group 
urls <- targets(lks)       # URLs of datasets in linkset
}
}
