% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R
\docType{methods}
\name{setPath}
\alias{setPath}
\alias{setPath,H5S_source,character-method}
\title{set path for hsds server resource}
\usage{
setPath(object, folderPath, ...)
}
\arguments{
\item{object}{H5S_source instance}

\item{folderPath}{character string with path to user's folder on hsds server}

\item{\dots}{not used}
}
\value{
an updated object with folderPath set
}
\description{
This function is deprecated and will be defunct in the next release.
}
\examples{
hsdsCon = H5S_source(URL_hsds()) # hsds server connection
setPath(hsdsCon, "/home/stvjc/hdf5_mat.h5")-> hsds
}
