% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R
\docType{methods}
\name{[,H5S_dataset,character,character,ANY-method}
\alias{[,H5S_dataset,character,character,ANY-method}
\alias{H5S_dataset,[,character,character-method}
\title{extract elements from H5S_dataset}
\usage{
\S4method{[}{H5S_dataset,character,character,ANY}(x, i, j, ...,
  drop = FALSE)
}
\arguments{
\item{x}{instance of H5S_dataset}

\item{i}{character string usable as select option for first matrix index in HDF5 server value API}

\item{j}{character string usable as select option for second matrix index in HDF5 server value API}

\item{\dots}{unused}

\item{drop}{logical defaults to FALSE}
}
\value{
matrix of data obtained
}
\description{
extract elements from H5S_dataset
}
